Move the question_report model to the general folder and change corresponding imports
